requirements.txt
flask
flask-cors
python-chess
openai
python-dotenv
